@@ -3,7 +3,7 @@ use std::convert::Infallible;
 use http::StatusCode;
 use ruma::{
     api::client::{
-        error::{Error as RumaError, ErrorKind},
+        error::{Error as RumaError, ErrorBody, ErrorKind},
         uiaa::{UiaaInfo, UiaaResponse},
     },
     OwnedServerName,
@@ -102,11 +102,14 @@ impl Error {
 
         if let Self::FederationError(origin, error) = self {
             let mut error = error.clone();
-            error.message = format!("Answer from {}: {}", origin, error.message);
+            error.body = ErrorBody::Standard {
+                kind: Unknown,
+                message: format!("Answer from {origin}: {error}"),
+            };
             return RumaResponse(UiaaResponse::MatrixError(error));
         }
 
-        let message = format!("{}", self);
+        let message = format!("{self}");
 
         use ErrorKind::*;
         let (kind, status_code) = match self {
@@ -131,8 +134,7 @@ impl Error {
         warn!("{}: {}", status_code, message);
 
         RumaResponse(UiaaResponse::MatrixError(RumaError {
-            kind,
-            message,
+            body: ErrorBody::Standard { kind, message },
             status_code,
         }))
     }
