@@ -63,6 +63,6 @@ pub use user_directory::*;
 pub use voip::*;
 
 pub const DEVICE_ID_LENGTH: usize = 10;
-pub const TOKEN_LENGTH: usize = 256;
-pub const SESSION_ID_LENGTH: usize = 256;
+pub const TOKEN_LENGTH: usize = 32;
+pub const SESSION_ID_LENGTH: usize = 32;
 pub const AUTO_GEN_PASSWORD_LENGTH: usize = 15;
