@@ -22,6 +22,10 @@ impl Service {
         self.db.highlight_count(user_id, room_id)
     }
 
+    pub fn last_notification_read(&self, user_id: &UserId, room_id: &RoomId) -> Result<u64> {
+        self.db.last_notification_read(user_id, room_id)
+    }
+
     pub fn associate_token_shortstatehash(
         &self,
         room_id: &RoomId,
@@ -36,10 +40,10 @@ impl Service {
         self.db.get_token_shortstatehash(room_id, token)
     }
 
-    pub fn get_shared_rooms<'a>(
-        &'a self,
+    pub fn get_shared_rooms(
+        &self,
         users: Vec<OwnedUserId>,
-    ) -> Result<impl Iterator<Item = Result<OwnedRoomId>> + 'a> {
+    ) -> Result<impl Iterator<Item = Result<OwnedRoomId>>> {
         self.db.get_shared_rooms(users)
     }
 }
